fix(cost): sub-cent precision + skip Twilio on non-Twilio calls
Four related issues surfaced after the rollup UI went live:

1. openai/twilio/embedding_cost_cents were unsigned int — Twilio
   inbound for a short RO call is 0.34¢, integer-truncated to 0.
   Columns (and their rollup equivalents) now numeric(12,4) /
   numeric(14,4); model casts + writer paths preserve float.
2. Negative duration_seconds on a handful of rows bled into the
   rollup ("-125.5 min"). Aggregator now uses
   GREATEST(duration_seconds, 0) so a broken row can't poison
   the tenant total.
3. Demo / test-vocal / web-only sessions had no metadata.provider
   but were still rolling up cost_cents with no breakdown. New
   command costs:backfill-breakdown recomputes openai from
   metadata.openai_usage and only assigns twilio when
   provider='twilio' — demo rows stay clean.
4. Re-running the daily aggregator didn't remove rows for bots
   that went idle. It now deletes the day's rollup slice first,
   so a re-run is a true replace (idempotent, but complete).

+++ app/Http/Controllers/Internal/MediaStreamEventController.php
@@ -223,8 +223,8 @@ class MediaStreamEventController extends Controller
             DB::transaction(function () use ($call, $openaiDelta) {
                 $call->refresh();
                 $call->update([
-                    'openai_cost_cents' => ((int) $call->openai_cost_cents) + $openaiDelta,
-                    'cost_cents'        => ((int) $call->cost_cents) + $openaiDelta,
+                    'openai_cost_cents' => (float) $call->openai_cost_cents + $openaiDelta,
+                    'cost_cents'        => (float) $call->cost_cents + $openaiDelta,
                 ]);
             });
         }
